refactor : unify class of oauth usecase #54
