+++ Trim-Api/src/main/java/trim/api/auth/controller/OAuthAccessApiController.java
@@ -4,11 +4,11 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-import trim.api.auth.service.google.GoogleSocialLoginUseCase;
-import trim.api.auth.service.google.GoogleSocialRegisterUserUseCase;
-import trim.api.auth.service.google.GoogleUserInfoUseCase;
-import trim.api.auth.service.kakao.KakaoUserInfoUseCase;
-import trim.api.auth.vo.GoogleRegisterUserRequest;
+import trim.api.auth.service.SocialLoginUseCase;
+import trim.api.auth.service.SocialRegisterUserUseCase;
+import trim.api.auth.service.GoogleUserInfoUseCase;
+import trim.api.auth.service.KakaoUserInfoUseCase;
+import trim.api.auth.vo.RegisterUserRequest;
 import trim.api.auth.vo.OauthUserInfoResponse;
 import trim.api.common.dto.ApiResponseDto;
 import trim.api.common.security.dto.JwtToken;
@@ -21,8 +21,8 @@ import trim.api.common.security.dto.JwtToken;
 public class OAuthAccessApiController {
 
     private final GoogleUserInfoUseCase googleUserInfoUseCase;
-    private final GoogleSocialLoginUseCase googleSocialLoginUseCase;
-    private final GoogleSocialRegisterUserUseCase googleSocialRegisterUserUseCase;
+    private final SocialLoginUseCase socialLoginUseCase;
+    private final SocialRegisterUserUseCase socialRegisterUserUseCase;
     private final KakaoUserInfoUseCase kakaoUserInfoUseCase;
 
     @GetMapping("/google/user-info")
@@ -30,19 +30,21 @@ public class OAuthAccessApiController {
         return ApiResponseDto.onSuccess(googleUserInfoUseCase.execute(code));
     }
 
-    @GetMapping("/google/login")
-    public ApiResponseDto<JwtToken> googleLogin(@RequestParam String email,
-                                                @RequestParam String socialType) {
-        return ApiResponseDto.onSuccess(googleSocialLoginUseCase.execute(email, socialType));
-    }
-
-    @PostMapping("/google/sign-up")
-    public ApiResponseDto<JwtToken> googleSignUp(@RequestBody GoogleRegisterUserRequest googleRegisterUserRequest) {
-        return ApiResponseDto.onSuccess(googleSocialRegisterUserUseCase.execute(googleRegisterUserRequest));
-    }
-
     @GetMapping("/kakao/user-info")
     public ApiResponseDto<OauthUserInfoResponse> kakaoUserInfo(@RequestParam("code") String code) {
         return ApiResponseDto.onSuccess(kakaoUserInfoUseCase.execute(code));
     }
+
+    @GetMapping("/login")
+    public ApiResponseDto<JwtToken> oauthLogin(@RequestParam String email,
+                                               @RequestParam String socialType) {
+        return ApiResponseDto.onSuccess(socialLoginUseCase.execute(email, socialType));
+    }
+
+    @PostMapping("/sign-up")
+    public ApiResponseDto<JwtToken> oauthSignUp(@RequestBody RegisterUserRequest registerUserRequest) {
+        return ApiResponseDto.onSuccess(socialRegisterUserUseCase.execute(registerUserRequest));
+    }
+
+
 }
