+++ Trim-Api/src/main/java/trim/api/auth/service/SocialRegisterUserUseCase.java
@@ -1,8 +1,8 @@
-package trim.api.auth.service.google;
+package trim.api.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
-import trim.api.auth.vo.GoogleRegisterUserRequest;
+import trim.api.auth.vo.RegisterUserRequest;
 import trim.api.common.security.dto.JwtToken;
 import trim.api.common.security.service.TokenService;
 import trim.common.annotation.UseCase;
@@ -17,16 +17,16 @@ import java.util.UUID;
 @UseCase
 @Transactional
 @RequiredArgsConstructor
-public class GoogleSocialRegisterUserUseCase {
+public class SocialRegisterUserUseCase {
 
     private final MemberDomainService memberDomainService;
     private final TokenService tokenService;
 
     //TODO nickname setting
-    public JwtToken execute(GoogleRegisterUserRequest googleRegisterUserRequest) {
+    public JwtToken execute(RegisterUserRequest registerUserRequest) {
         Profile profile = Profile.builder()
-                .email(googleRegisterUserRequest.getEmail())
-                .socialType(EnumConvertUtil.convert(SocialType.class, googleRegisterUserRequest.getOauthProvider()))
+                .email(registerUserRequest.getEmail())
+                .socialType(EnumConvertUtil.convert(SocialType.class, registerUserRequest.getOauthProvider()))
                 .username(UUID.randomUUID().toString())
                 .build();
         memberDomainService.registerMember(profile, Role.USER, UUID.randomUUID().toString());
