--- Trim-Api/src/main/java/trim/api/auth/vo/KakaoUserInfoResponse.java
@@ -1,4 +1,0 @@
-package trim.api.auth.vo;
-
-public class KakaoUserInfoResponse {
-}
