+++ Trim-Api/src/main/java/trim/api/auth/vo/RegisterUserRequest.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class GoogleRegisterUserRequest {
+public class RegisterUserRequest {
     private final String email;
     private final String oauthProvider;
 }
